Technical/Change RSpec helpers namespace
