+++ .reek.yml
@@ -11,35 +11,35 @@ detectors:
   UtilityFunction:
     exclude:
       - DnsMock::Server::RecordsDictionaryBuilder#build_records_instances_by_type
-      - DnsMock::RecordsDictionaryHelper#hostname_records_by_type
-      - DnsMock::DnsMessageHelper#to_dns_name
-      - DnsMock::ContextGeneratorHelper#random_dns_record_type
-      - DnsMock::ContextGeneratorHelper#random_hostname
-      - DnsMock::ContextGeneratorHelper#create_dns_name
-      - DnsMock::ContextGeneratorHelper#random_ip_v4_address
-      - DnsMock::ContextGeneratorHelper#random_ip_v6_address
-      - DnsMock::ContextGeneratorHelper#random_txt_record_context
-      - DnsMock::ContextGeneratorHelper#to_punycode_hostname
-      - DnsMock::RecordsDictionaryHelper#create_records_dictionary_by_records
-      - DnsMock::ServerHelper#start_random_server
-      - DnsMock::ServerHelper#stop_all_running_servers
-      - DnsMock::ContextGeneratorHelper#random_port_number
+      - DnsMock::RspecHelper::RecordsDictionary#hostname_records_by_type
+      - DnsMock::RspecHelper::DnsMessage#to_dns_name
+      - DnsMock::RspecHelper::ContextGenerator#random_dns_record_type
+      - DnsMock::RspecHelper::ContextGenerator#random_hostname
+      - DnsMock::RspecHelper::ContextGenerator#create_dns_name
+      - DnsMock::RspecHelper::ContextGenerator#random_ip_v4_address
+      - DnsMock::RspecHelper::ContextGenerator#random_ip_v6_address
+      - DnsMock::RspecHelper::ContextGenerator#random_txt_record_context
+      - DnsMock::RspecHelper::ContextGenerator#to_punycode_hostname
+      - DnsMock::RspecHelper::RecordsDictionary#create_records_dictionary_by_records
+      - DnsMock::RspecHelper::Server#start_random_server
+      - DnsMock::RspecHelper::Server#stop_all_running_servers
+      - DnsMock::RspecHelper::ContextGenerator#random_port_number
       - DnsMock::TestFramework::RSpec::Helper#dns_mock_server
-      - DnsMock::ContextGeneratorHelper#to_rdns_hostaddress
+      - DnsMock::RspecHelper::ContextGenerator#to_rdns_hostaddress
 
   DuplicateMethodCall:
     exclude:
-      - DnsMock::ContextGeneratorHelper#create_records
+      - DnsMock::RspecHelper::ContextGenerator#create_records
 
   ControlParameter:
     exclude:
       - DnsMock::Error::Helper#raise_unless
       - DnsMock::Server#initialize
-      - DnsMock::ContextGeneratorHelper#to_punycode_hostname
+      - DnsMock::RspecHelper::ContextGenerator#to_punycode_hostname
 
   MissingSafeMethod:
     exclude:
-      - DnsMock::PortInUseHelper::TransportService
+      - DnsMock::RspecHelper::PortInUse::TransportService
       - DnsMock::Server
 
   TooManyStatements:
